move histogram to internal.metrics
