+++ sdks/python/apache_beam/metrics/cells_test.py
@@ -28,11 +28,6 @@ from apache_beam.metrics.cells import DistributionCell
 from apache_beam.metrics.cells import DistributionData
 from apache_beam.metrics.cells import GaugeCell
 from apache_beam.metrics.cells import GaugeData
-from apache_beam.metrics.cells import HistogramCell
-from apache_beam.metrics.cells import HistogramCellFactory
-from apache_beam.metrics.cells import HistogramData
-from apache_beam.utils.histogram import Histogram
-from apache_beam.utils.histogram import LinearBucket
 
 
 class TestCounterCell(unittest.TestCase):
@@ -150,52 +145,6 @@ class TestGaugeCell(unittest.TestCase):
     # the final result.
     result = g2.combine(g1)
     self.assertEqual(result.data.value, 1)
-
-
-class TestHistogramCell(unittest.TestCase):
-  @classmethod
-  def _modify_histogram(cls, d):
-    for i in range(cls.NUM_ITERATIONS):
-      d.update(i)
-
-  NUM_THREADS = 5
-  NUM_ITERATIONS = 100
-
-  def test_parallel_access(self):
-    # We create NUM_THREADS threads that concurrently modify the distribution.
-    threads = []
-    bucket_type = LinearBucket(0, 1, 100)
-    d = HistogramCell(bucket_type)
-    for _ in range(TestHistogramCell.NUM_THREADS):
-      t = threading.Thread(
-          target=TestHistogramCell._modify_histogram, args=(d, ))
-      threads.append(t)
-      t.start()
-
-    for t in threads:
-      t.join()
-
-    histogram = Histogram(bucket_type)
-    for _ in range(self.NUM_THREADS):
-      for i in range(self.NUM_ITERATIONS):
-        histogram.record(i)
-
-    self.assertEqual(d.get_cumulative(), HistogramData(histogram))
-
-  def test_basic_operations(self):
-    d = HistogramCellFactory(LinearBucket(0, 1, 10))()
-    d.update(10)
-    self.assertEqual(
-        str(d.get_cumulative()),
-        'HistogramData(Total count: 1, P99: >=10, P90: >=10, P50: >=10)')
-    d.update(0)
-    self.assertEqual(
-        str(d.get_cumulative()),
-        'HistogramData(Total count: 2, P99: >=10, P90: >=10, P50: 1)')
-    d.update(5)
-    self.assertEqual(
-        str(d.get_cumulative()),
-        'HistogramData(Total count: 3, P99: >=10, P90: >=10, P50: 6)')
 
 
 if __name__ == '__main__':
